Abstract metrics from reviewers/interviewers.
This commit abstracts the individual social, technical, etc. metrics
from the reviewer and interviewer tables into a separate table,
"metric", which stores the metric name and whether or not it is
currently active, i.e. to be included in newly created application
review records or interview records. The tables "reviewer_metric" and
"interviewer_metric" are introduced to link a reviewer's score to the
respective metric and review being conductive; likewise with interviews.