Prototype frontend infrastructure
I still need to put together some examples and documentation on making
and editing pages. The infrastructure might change as I put that
together, depending on how easy that ends up being.

The goal of the frontend infrastructure is first to be as easy as
possible for anybody to add and change pages for the frontend, and
second to be readable and laid out sanely.