Initial Test Suite build out
tasty has been chosen as the test framework to orchestrate our tests.
It bears similarities to test-framework; however, it has a better
design, and it is actively maintained.

doctest is to be run as a unit test for every module. This will verify
code examples and properties in our documentation, and I expect this to
be used heavily.

HUnit will be used for unit testing. Shocking.

Quickcheck is going to be used for property based tests.