Rewrite all statements with the HaSQL QuasiQuoter.
What it says on the tin. Now all of our statements are guaranteed to be
typesafe. Strangely, all enumerated types are passed in as a Data.Text
values. This is to mirror the native type mapping used by HaSQL's
Postgres driver. Postgres expects any data type coming in as an
enumerated type to be parseable as a string. Higher level library
functions will of  unwrap an ADT for these parameters; I felt it
was inappropriate to put that machinery at this level in the system.

I prettied up the docs a bit, but I need to go back in and document the
query parameters because Haddock can now be made aware of them
explicitly.

Also, how do I write tests for this? I feel like that needs a sizeable
corpus of test data.